requirements.txt
torch==2.0.1
Pillow==10.0.0
torchvision==0.15.2
streamlit
altair<5
scipy==1.11.1
pysolr==3.9.0